pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "arc-solver"
version = "0.1.0"
description = "ARC-AGI Solver with mathematical perception and DSL reasoning"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "arc_solver", from = "src"}]

[tool.poetry.scripts]
arc-solver = "arc_solver.cli.main:main"

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.24.0"
scipy = "^1.8.0"
scikit-image = "^0.20.0"
numba = "^0.57.0"
pybind11 = "^2.11.0"
redis = "^4.5.0"
hydra-core = "^1.3.0"
gudhi = "^3.8.0"
pytest = "^7.4.0"
torch = {version = "^2.0.0", optional = true}
transformers = {version = "^4.30.0", optional = true}
bitsandbytes = {version = "^0.41.0", optional = true}
accelerate = {version = "^0.20.0", optional = true}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"
pre-commit = "^3.4.0"

[tool.poetry.group.gpu.dependencies]
cupy = {version = "^12.0.0", optional = true}

[tool.poetry.extras]
gpu = ["cupy"]
llm = ["torch", "transformers", "bitsandbytes", "accelerate"]
all = ["cupy", "torch", "transformers", "bitsandbytes", "accelerate"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true